drf/dev/requirements.txt
Django==4.0.6
djangorestframework
pyyaml
requests
django-cors-headers
dj-rest-auth
django-allauth